Initial work on wireless bootloading logic
The PROG_MULTI loop now sends an address to program, the length, the data,
a CRC of the data and an EOC. If the address is unexpected an incorrect
address message is sent back. If the CRC is incorrect, a bad CRC message
is sent back.

When a "bad message" resonse is sent back the input buffer is flushed, so that
it is empty for the next message that arrives. This prevents errors where
the remaining data in the buffer could be interpreted as a command to which
the bootloader tries to respond.