Adds auto dependencies for make (#85)
* Adds auto dependencies for make

This resolves the issue that if any of the header header files were
modified, make would not recompile to include the changes in the header
files the unless clean was run before recompiling.

* Moves common build commands into common rules.mk makefile

The idividual makefiles all have the same makefile structure. To prevent
duplicating the same commands in each makefile, the common commands have
been broken out into their own dedicated rules.mk makefile which is included
in each of the individual makefiles.

* Adds individual build directories for each bootloader

This resolves the issue where the dependencies for .o files
would be met and the object file is not rebuilt for a new target,
even though the build flags might have changed; this is the case
when compiling for an F4 board compared to an F1 board.

* Fixes compiling in build directory with auto dependencies

This commit solves the following issues related to why the executable would
continue to be rebuild even when no changes to files had been made. The reasons
for this are as follows:
 - When the object files were created they did not use the dependency files
   (the -M or -MM option)
 - The targets for the object files and ELF and binary files did not exist,
   so the object target would always be rebuilt.

* Fixes 'make deploy' for bin files now being generated in individual build folders

* Adds sizes target

Prints the size of the text, data, and bss of the bootloader firmware.

* Makefile clean: bug fix to avoid error when running 'make clean' twice consecutively

On the second invocation of 'make clean' or on the first when no build directories
exist yet, the following error will be produced
    rm: cannot remove ‘build_*’: No such file or directory
    make: *** [clean] Error 1

To avoid this, we need to envoke 'rm' with the '-rf' option instead of the '-r' option.

* Cut down on noise

* Makefile: to be consistent, uses MAKE expansion define for building libopencm3 as well

* Makefile: Fixes 'sizes' target when building with existing elf files in Bootloader directory

When the sizes target is run in the Bootloader directory and there are existing
.elf files that are present in the Bootloader directory (not in a build directory)
then one gets the following error:
   arm-none-eabi-objcopy -O binary <build_directory>/<elf_filename>.elf <build_directory>/<binary_filename>.bin
   find: paths must precede expression: <elf_filename>.elf
   Usage: find [-H] [-L] [-P] [-Olevel] [-D help|tree|search|stat|rates|opt|exec|time] [path...] [expression]

This commit fixes this issue.

* Makefile: clean now removes bin and elf files in directly in Bootloader directory as well

* Makefile: adds explanatory comments to clean commands