Restructures USB VBUS check to single function
The initilaisation, reading and deinitiliasation of PA9 is
preformed all in board_test_usb_vbus().

This also solves the problem, where PA9 was floating and
could erroneously have been read as being HIGH, and hence triggered
the board to stay in the bootloader until timeout, instead of booting
immediately.

Fixes issue #82.