Change USBPRODUCTID for TARGET_HW_OMNIBUSF4SD
There are issues with connecting FCUs based on omnibusf4sd to the latest version of QGroundControl. This is because QGroundControl expects omnibusf4sd to be of productID 0x0001 (as seen in [qgroundcontrol/src/comm/USBBoardInfo.json](https://github.com/mavlink/qgroundcontrol/blob/master/src/comm/USBBoardInfo.json)) but the bootloader reports it as productID 0x0016. This commit aims to fix this discrepancy and restore the productID value to the official one